feat: Add renderer/volume quality presets and fixes
Adds renderer-feature and volume quality presets, shader precision and build compatibility fixes, and several runtime/editor optimizations.

Key changes:
- Added Renderer Feature quality preset (Manual/Low/Medium/High/Cinematic) controlling resolution scale and upscale mode with apply-on-change behavior (VolumetricCloudsURP.cs).
- Added Volume quality preset (Manual/Low/Medium/High/Cinematic) that sets raymarch parameters (primary/light steps, temporal accumulation, perceptual blending) with apply-on-change behavior (VolumetricCloudsVolume.cs).
- Cached local material keyword toggles to avoid reapplying keywords every frame and added a SetKeyword helper to minimize shader keyword churn (VolumetricCloudsURP.cs).
- Shader precision cleanup: changed _NumPrimarySteps, _NumLightSteps and _MaxStepSize from half to float (VolumetricCloudsDefs.hlsl) for better precision/compatibility.
- Unity 6.5 compatibility: guarded removed non-RenderGraph OnCameraSetup/Execute overrides and documented RenderGraph as primary path (CHANGELOG.md/README.md).
- Fixed runtime issues: removed per-frame weather-preset reapplication, gated custom LUT rebuild/upload, and added null guards for main light and CloudsMaterial setter.
- Editor/UI: expose Volume quality preset in the volume inspector (VolumetricCloudsVolumeEditor.cs).
- README and CHANGELOG updated with installation, setup, preset documentation, known limitations and testing notes.
- Sample material tweaks: updated valid keywords and adjusted several default material property values (Samples/VolumetricClouds.mat).

Files modified: CHANGELOG.md, README.md, Editor/VolumetricCloudsVolumeEditor.cs, Runtime/VolumetricCloudsDefs.hlsl, Runtime/VolumetricCloudsURP.cs, Runtime/VolumetricCloudsVolume.cs, Samples/VolumetricClouds.mat.

+++ Runtime/VolumetricCloudsDefs.hlsl
@@ -3,9 +3,9 @@
 
 CBUFFER_START(UnityPerMaterial)
 float _Seed;
-half _NumPrimarySteps;
-half _NumLightSteps;
-half _MaxStepSize;
+float _NumPrimarySteps;
+float _NumLightSteps;
+float _MaxStepSize;
 float _HighestCloudAltitude;
 float _LowestCloudAltitude;
 half4 _ShapeNoiseOffset;
